Resolving conflict in login and signup page
